requirements.txt
Flask==3.1.3
langchain_community==0.4.1
langchain_core==1.2.18
langchain_groq==1.1.2
langchain_huggingface==1.2.1
langchain_pinecone==0.2.13
langchain_text_splitters==1.1.1
pinecone>=5.0.0
python-dotenv==1.2.2
rank_bm25==0.2.2
scikit-learn>=1.4.0
Werkzeug==3.1.6
